+++ CHANGES.rst
@@ -3,6 +3,17 @@ CHANGES
 =======
 
 ---
+1.6
+---
+
+*  GNOME Keyring backend:
+
+   - Use the same attributes (``username`` / ``service``) as the SecretService
+     backend uses, allow searching for old ones for compatibility.
+   - Also set ``application`` attribute.
+   - Correctly handle all types of errors, not only ``CANCELLED`` and ``NO_MATCH``.
+
+---
 1.5
 ---